refactor project modules
BREAKING CHANGE: changed whole src structure
