+++ src/management.py
@@ -1,10 +1,10 @@
-from flask import Flask, Response, jsonify, request
+from flask import Response, jsonify, request
 from mininet.net import Mininet
+from typing import List
 import requests
 
-from qos import FlowManager
-from information import Utils
-from topo import Validator, Topology, Controller
+from utils import Api
+from data import Policy
 
 
 class ActionDirector:
@@ -76,12 +76,9 @@ class ActionDirector:
     def show_manual() -> None:
         pass
 
-class Api:
-    app = Flask(__name__)
-
 class ServicesController:
     def __init__(self, topology_filepath: str):
-        self.__config = Utils.read_json_from() # ler do caminho do projeto
+        self.__config = File.read_json_from() # ler do caminho do projeto
         self.__topo_manager = TopoManager(topology_filepath=topology_filepath)
         self.__flow_manager = FlowManager()
 
@@ -172,3 +169,76 @@ class TopoManager:
     def report_topology_state(self) -> bool:
         return True
 
+class FlowManager:
+    def __init__(self):
+        self.__config: dict = {}
+        self.policies: List[Policy] = []
+
+    def __validate(self, policy: Policy) -> RoutineResults:
+        try:
+            errors = []
+
+            if not policy.name:
+                errors.append("Nome de política inválido.")
+            
+            if not isinstance(policy.traffic_type, Policy.PolicyTypes):
+                errors.append("Tipo de tráfego fornecido é inválido.")
+            
+            if not (0 < policy.bandwidth_reserved <= 1000):
+                errors.append("Largura de banda reservada deve estar entre 1 e 100 Mbps.")
+
+            if errors:
+                return RoutineResults(status=False, err_reason="; ".join(errors))
+            
+        except Exception as e:
+            return RoutineResults(status=False, err_reason="Erro de validação: {str(e)}")
+        
+        finally:
+            print("Política validada com sucesso.")
+
+    def __init_framework_config(self) -> RoutineResults:
+        # inicializa o arquivo de config do controlador
+        pass
+
+    def __update_tables(self) -> RoutineResults:
+        # altera o arquivo de config do controlador para lidar
+        # com uma nova política ou com redirecionamento de tráfego
+        pass
+
+    def __process_alerts(self) -> RoutineResults:
+        # recebe o alerta do monitor
+        # chama redirect_traffic se necessário
+        pass
+
+    def redirect_traffic(self) -> RoutineResults:
+        pass
+
+    def create(self, policy: Policy) -> RoutineResults:
+        try:
+            validation = self.__validate(policy)
+
+            if not validation.status == False:
+                return validation
+            
+            self.policies.append(policy)
+
+            update = self.__update_tables()
+            
+            if update.status == False:
+                self.policies.remove(policy)
+                return update
+            
+            return RoutineResults(status=True, payload="Política criada com sucesso.")
+        
+        except Exception as e:
+            print(f"Erro ao criar política: {e}")
+            return RoutineResults(status=False, err_reason=str(e))
+        
+        finally:
+            print("Operação de criação de política finalizada.")
+
+    def update(self, policy: Policy) -> RoutineResults:
+        pass
+
+    def remove(self, policy: Policy) -> RoutineResults:
+        pass
